feat: add compat option skipConfigurationInfoQuery
